+++ Cargo.toml
@@ -1,7 +1,15 @@
 [workspace]
-members = ["compio", "compio-buf", "compio-driver", "compio-macros"]
+members = [
+    "compio",
+    "compio-buf",
+    "compio-driver",
+    "compio-runtime",
+    "compio-macros",
+]
 resolver = "2"
 
 [workspace.dependencies]
 compio-buf = { path = "./compio-buf" }
+compio-driver = { path = "./compio-driver" }
+compio-runtime = { path = "./compio-runtime" }
 compio-macros = { path = "./compio-macros" }
