+++ compio-driver/Cargo.toml
@@ -9,14 +9,14 @@ compio-buf = { workspace = true }
 
 # Utils
 cfg-if = "1"
-# may be excluded from linking if the unstable equivalent is used
-once_cell = "1"
 slab = "0.4"
 socket2 = { version = "0.5", features = ["all"] }
 
 # Windows specific dependencies
 [target.'cfg(target_os = "windows")'.dependencies]
 compio-buf = { workspace = true, features = ["arrayvec"] }
+# may be excluded from linking if the unstable equivalent is used
+once_cell = "1"
 widestring = "1"
 windows-sys = { version = "0.48", features = [
     "Win32_Foundation",
@@ -43,3 +43,7 @@ libc = "0.2"
 
 [features]
 default = ["io-uring"]
+
+# Nightly features
+once_cell_try = []
+nightly = ["once_cell_try"]
